base_wamas_ubl: move lib tests to lib and run them all
