+++ base_wamas_ubl/lib/test_wamas2dict
@@ -1,0 +1,2 @@
+#!/usr/bin/sh
+/usr/bin/env python3 -m wamas.tests.test_wamas2dict
