+++ base_wamas_ubl/lib/wamas/dict2wamas.py
@@ -5,6 +5,7 @@ import argparse
 import logging
 from ast import literal_eval
 from pprint import pformat
+
 from freezegun import freeze_time
 
 from . import const, utils
