+++ base_wamas_ubl/lib/wamas/tests/samples/UBL2WAMAS-SAMPLE_AUSK_AUSP-DESPATCH_ADVICE2.xml
