+++ base_wamas_ubl/lib/wamas/tests/samples/UBL2WAMAS-SAMPLE_KRETK_KRETP-DESPATCH_ADVICE.xml
