+++ base_wamas_ubl/lib/wamas/tests/samples/UBL2WAMAS-SAMPLE_KRETK_KRETP.wamas
