+++ base_wamas_ubl/lib/wamas/tests/samples/UBL2WAMAS-SAMPLE_WEAK_WEAP-DESPATCH_ADVICE.xml
