+++ base_wamas_ubl/lib/wamas/tests/samples/WAMAS2UBL-SAMPLE_AUSKQ_WATEKQ_WATEPQ.wamas
