+++ base_wamas_ubl/lib/wamas/tests/samples/WAMAS2UBL-SAMPLE_KRETKQ_KRETPQ-DESPATCH_ADVICE.xml
