+++ base_wamas_ubl/lib/wamas/tests/samples/WAMAS2UBL-SAMPLE_KRETKQ_KRETPQ.wamas
