+++ base_wamas_ubl/lib/wamas/tests/test_ubl2wamas.py
@@ -1,7 +1,10 @@
 import unittest
 from datetime import date, datetime
 
-from ..utils import set_value_to_string
+from freezegun import freeze_time
+
+from ..ubl2wamas import ubl2wamas
+from ..utils import file_open, file_path, set_value_to_string
 
 
 class TestUbl2wamas(unittest.TestCase):
@@ -101,6 +104,37 @@ class TestUbl2wamas(unittest.TestCase):
                     input_val, ttype, length, dp, do_convert_tz=True
                 )
                 self.assertEqual(output_val, expected_output_val)
+
+    @freeze_time("2023-05-01")
+    def _convert_ubl2wamas(
+        self, input_filename, expected_output_filename, telegram_type
+    ):
+        path = file_path("tests/samples/")
+        with file_open(path + input_filename) as inputfile, file_open(
+            path + expected_output_filename
+        ) as outputfile:
+            str_input = inputfile.read()
+            output = ubl2wamas(str_input, telegram_type)
+            expected_output = outputfile.read().strip("\n")
+            self.assertEqual(output, expected_output)
+
+    def test_convert_ubl2wamas_picking(self):
+        input_file = "UBL2WAMAS-SAMPLE_AUSK_AUSP-DESPATCH_ADVICE2.xml"
+        expected_output = "UBL2WAMAS-SAMPLE_AUSK_AUSP.wamas"
+        msg_type = "Picking"
+        self._convert_ubl2wamas(input_file, expected_output, msg_type)
+
+    def test_convert_ubl2wamas_reception(self):
+        input_file = "UBL2WAMAS-SAMPLE_WEAK_WEAP-DESPATCH_ADVICE.xml"
+        expected_output = "UBL2WAMAS-SAMPLE_WEAK_WEAP.wamas"
+        msg_type = "Reception"
+        self._convert_ubl2wamas(input_file, expected_output, msg_type)
+
+    def test_convert_ubl2wamas_return(self):
+        input_file = "UBL2WAMAS-SAMPLE_KRETK_KRETP-DESPATCH_ADVICE.xml"
+        expected_output = "UBL2WAMAS-SAMPLE_KRETK_KRETP.wamas"
+        msg_type = "Return"
+        self._convert_ubl2wamas(input_file, expected_output, msg_type)
 
 
 if __name__ == "__main__":
